backend/pipeline/requirements.txt

google-cloud-bigquery
google-cloud-firestore
functions-framework
cloudevents

